requirements.txt
python-dotenv
SQLAlchemy==1.4.46
aiohttp
asyncpg
more_itertools